// FE/src/api/generated/models/orderResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * awr-store-backend-project
 * OpenAPI spec version: 1.0.0
 */
import type { OrderResponseStatus } from "./orderResponseStatus";
import type { OrderResponseProductsItem } from "./orderResponseProductsItem";

export interface OrderResponse {
  /** Order ID */
  id: string;
  /** Customer ID */
  customerId: string;
  /** Order creation timestamp */
  orderCreatedDate: unknown;
  /** Order last updated timestamp */
  orderUpdatedDate: unknown;
  /** Order status */
  status: OrderResponseStatus;
  /** Total order amount */
  orderTotal: number;
  /** Ordered products */
  products: OrderResponseProductsItem[];
}